Starting popup (probably broke something)
Tried to implement popup with Toplevel() function, however likely? ran into issues with the order in which functions ran. I am assuming this is because it is managed by the window manager and therefore skips for some reason at runtime, but could be wrong. Definitely broken, going to try a normal tkinter window tomorrow and see of that solves the problem. Also ran into issues with dynamic labels and progress bars. All on the to-do list.